# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smoothy King API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /user/onboarding:
    post:
      operationId: submitOnboarding
      tags: [user]
      summary: Save onboarding profile data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingInput"
      responses:
        "200":
          description: Onboarding saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfileResponse"
        "401":
          description: Not authenticated
  /user/profile:
    get:
      operationId: getUserProfile
      tags: [user]
      summary: Get the current user's onboarding profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfileResponse"
        "401":
          description: Not authenticated
        "404":
          description: Profile not found
    put:
      operationId: updateUserProfile
      tags: [user]
      summary: Update the current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingInput"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfileResponse"
        "401":
          description: Not authenticated
  /recipes:
    get:
      operationId: listRecipes
      tags: [recipes]
      summary: List all recipes
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: benefit
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recipe"
  /recipes/{id}:
    get:
      operationId: getRecipe
      tags: [recipes]
      summary: Get a single recipe by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recipe found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
        "404":
          description: Not found
...
components:
  schemas:
...
    Recipe:
      type: object
      required: [id, name, slug, category, tagline, prepTimeMinutes, servings, benefits, ingredients, steps, imageUrl, isFeatured]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        category:
          type: string
          description: "smoothie | shaker | meal"
        tagline:
          type: string
          description: "Short poetic descriptor (e.g. 'The Glow Ritual')"
        description:
          type: ["string", "null"]
        prepTimeMinutes:
          type: integer
        servings:
          type: integer
        calories:
          type: ["integer", "null"]
        protein:
          type: ["number", "null"]
        benefits:
          type: array
          items:
            type: string
          description: "e.g. ['glowy-skin', 'hydration', 'anti-inflammatory']"
        skinBenefitScore:
          type: ["integer", "null"]
          description: "1-10 skin benefit rating inspired by K-beauty principles"
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/RecipeIngredient"
        steps:
          type: array
          items:
            type: string
        imageUrl:
          type: string
        isFeatured:
          type: boolean
        difficulty:
          type: ["string", "null"]
          description: "easy | medium | advanced"
        tags:
          type: array
          items:
            type: string
    RecipeIngredient:
      type: object
      required: [name, amount, unit]
      properties:
        name:
          type: string
        amount:
          type: string
        unit:
          type: string
        benefit:
          type: ["string", "null"]
          description: "Primary skin/health benefit of this ingredient"
...
    OnboardingInput:
      type: object
      properties:
        gender:
          type: string
          description: "male | female | other"
        age:
          type: integer
        height:
          type: number
          description: "cm"
        weight:
          type: number
          description: "kg"
        activityLevel:
          type: string
          description: "sedentary | light | moderate | active | very_active"
        allergies:
          type: array
          items:
            type: string
        dislikedIngredients:
          type: array
          items:
            type: string
        primaryGoal:
          type: string
          description: "glowy-skin | hydration | sun-ritual | protein-power | anti-inflammatory | detox-clarity"
        secondaryGoals:
          type: array
          items:
            type: string
        tastePreference:
          type: array
          items:
            type: string
          description: "sweet | sour | nutty | fresh"
    UserProfileResponse:
      type: object
      required: [id, userId]
      properties:
        id:
          type: integer
        userId:
          type: integer
        gender:
          type: ["string", "null"]
        age:
          type: ["integer", "null"]
        height:
          type: ["number", "null"]
        weight:
          type: ["number", "null"]
        activityLevel:
          type: ["string", "null"]
        allergies:
          type: array
          items:
            type: string
        dislikedIngredients:
          type: array
          items:
            type: string
        primaryGoal:
          type: ["string", "null"]
        secondaryGoals:
          type: array
          items:
            type: string
        tastePreference:
          type: array
          items:
            type: string
        createdAt:
          type: string
        updatedAt:
          type: string