# lib/api-zod/src/generated/types/onboardingInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Smoothy King API
 * OpenAPI spec version: 0.1.0
 */

export interface OnboardingInput {
  /** male | female | other */
  gender?: string;
  age?: number;
  /** cm */
  height?: number;
  /** kg */
  weight?: number;
  /** sedentary | light | moderate | active | very_active */
  activityLevel?: string;
  allergies?: string[];
  dislikedIngredients?: string[];
  /** glowy-skin | hydration | sun-ritual | protein-power | anti-inflammatory | detox-clarity */
  primaryGoal?: string;
  secondaryGoals?: string[];
  /** sweet | sour | nutty | fresh */
  tastePreference?: string[];
}